+++ Changelog.md
@@ -30,6 +30,7 @@
 - Changed model validation errors to use built-in error key resolution instead of inline `I18n.t` calls (#7806)
 - Upgraded to Rails v8.1 (#7815)
 - Replaced MathJax with KateX (#7847)
+- Updated the student table to use `@tanstack/react-table` v8 (#7826)
 
 ## [v2.9.4]